add service

+++ Crypto_App/lib/constants/path/api_path.dart
@@ -1,5 +1,5 @@
 class ApiPath {
-  String cryptoConvert({required String primaryCur, required String secondCur, required int amount}) {
+  String cryptoConvert({required String primaryCur, required String secondCur, required double amount}) {
     return ("https://api.coinconvert.net/convert/$primaryCur/$secondCur?amount=$amount");
   }
 }
